Starting work with Electron
- Added `renderer.js` from Electron's starter app
- Changes to make the Electron app start
- Moved fetching code to `src/fetch.ts`
- Included `fetch` module from the `main`

TODO: change the view to display the fetched data
