+++ package.json
@@ -17,7 +17,7 @@
     "typescript": "^3.2.1"
   },
   "scripts": {
-    "start": "node dist/main.js",
+    "start": "electron ./dist/main",
     "build": "npx tsc",
     "test": "npx jest --verbose",
     "test-watch": "npx jest --verbose --watchAll"
